feat: revert banner
