+++ src/lib/constants.ts
@@ -1,5 +1,5 @@
 export const GITHUB_STARS = '42.8K';
-export const BANNER_KEY = 'init-banner-02'; // Change key to force banner to show again
+export const BANNER_KEY: Banners = 'discord-banner-01'; // Change key to force banner to show again
 export const SENTRY_DSN =
     'https://27d41dc8bb67b596f137924ab8599e59@o1063647.ingest.us.sentry.io/4507497727000576';
 
@@ -9,6 +9,8 @@ export const SENTRY_DSN =
  * init-banner-02
  * pricing-banner-01
  */
+type Banners = 'discord-banner-01' | 'init-banner-02' | 'pricing-banner-01'
+
 
 export type Social = {
     icon: string;
